Added connection checks in UpdateService.
Imported the EuroGen.Watcher namespace to monitor Internet connectivity.

Added checks to ensure that the application has Internet access and that the GitHub site is available before fetching updates.

Modified the handling of the current version to no longer remove the 'v' prefix when comparing versions.

Changed exception handling during version comparison to return false in case of a parsing error.